requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
